fix: 13022 is a S16 in unitary watts for SH5.0RS
This change ensures that my unit returns sensible readings which match
the official app's displayed data as of today. The change to a S16
contradicts the 1.0.20 and 1.0.23 communications protocol documents
which states the register is a U16, but power flow being signed makes
more sense in any case. I haven't been able to find a more recently
published document to confirm if Sungrow have updated this since the
1.0.23 document was published, possibly as the result of a change to the
register definition as part of a firmware update since released.